Move walletgen functions into lib
