+++ src/bridge/views/InviteWallet.js
@@ -15,19 +15,10 @@ import { BRIDGE_ERROR } from '../lib/error'
 import { Row, Col, Warning, Button, H1, H3 } from '../components/Base'
 
 // for wallet generation
-import * as kg from '../../../node_modules/urbit-key-generation/dist/index'
-import * as more from 'more-entropy'
-import lodash from 'lodash'
+import * as wg from '../../walletgen/lib/lib'
 import PaperCollateralRenderer from 'PaperCollateralRenderer'
 import JSZip from 'jszip'
 import saveAs from 'file-saver'
-import { MAX_GALAXY, MIN_STAR, MAX_STAR, MIN_PLANET,
-         GALAXY_ENTROPY_BITS, STAR_ENTROPY_BITS, PLANET_ENTROPY_BITS,
-         SEED_ENTROPY_BITS,
-         GEN_STATES
-       } from '../../walletgen/lib/constants'
-const SEED_LENGTH_BYTES = SEED_ENTROPY_BITS / 8
-const NEXT_STEP_NUM = 6;
 
 class InviteWallet extends React.Component {
 
@@ -74,63 +65,10 @@ class InviteWallet extends React.Component {
     //
   }
 
-  //TODO pulled from walletgen/views/Generate and Download, put into lib
   async generateWallet(point) {
-    const makeTicket = point => {
-
-      const bits = point < MIN_STAR
-        ? GALAXY_ENTROPY_BITS
-        : point < MIN_PLANET
-          ? STAR_ENTROPY_BITS
-          : PLANET_ENTROPY_BITS
-
-      const bytes = bits / 8
-      const some = new Uint8Array(bytes)
-      window.crypto.getRandomValues(some)
-
-      const gen = new more.Generator()
-
-      return new Promise((resolve, reject) => {
-        gen.generate(bits, result => {
-          const chunked = lodash.chunk(result, 2)
-          const desired = chunked.slice(0, bytes) // only take required entropy
-          const more = lodash.flatMap(desired, arr => arr[0] ^ arr[1])
-          const entropy = lodash.zipWith(some, more, (x, y) => x ^ y)
-          const buf = Buffer.from(entropy)
-          const patq = ob.hex2patq(buf.toString('hex'))
-          resolve(patq)
-          reject('Entropy generation failed')
-        })
-      })
-    }
-
-    const genWallet = async (point, ticket, cb) => {
-
-      const config = {
-        ticket: ticket,
-        seedSize: SEED_LENGTH_BYTES,
-        ship: point,
-        password: '',
-        revisions: {},
-        boot: false //TODO should this generate networking keys here already?
-      };
-
-      const wallet = await kg.generateWallet(config);
-
-      // This is here to notify the anyone who opens console because the thread
-      // hangs, blocking UI updates so this cannot be doen in the UI
-      console.log('Generating Wallet for point address: ', point);
-
-      return wallet;
-    }
-
-    const ticket = await makeTicket(point);
-
-    const wallet = await genWallet(point, ticket);
-
+    const ticket = await wg.makeTicket(point);
+    const wallet = await wg.generateWallet(point, ticket);
     this.setState({ wallet: Just(wallet) });
-    console.log('got wallet', wallet);
-    console.log('should start rendering now');
   }
 
   download() {
